only call zoomeye if user given username pasword for it
Lots of times zoomeye api is down or not responding intime so the whole main program get hang waiting for zoomeye response. There was no option for user in configuration to not to use zoomeye. So having a simple check for empty username/password helps the user. If someone dont want to use zoomeye he can empty the username/password.